rename spec to component
