+++ internal/docs/component/copy-button.js
