+++ internal/docs/component/default-props.js
