+++ internal/docs/component/example.js
