+++ internal/docs/component/get-defaults-from-code.js
