+++ internal/docs/component/icon-browser.js
