+++ internal/docs/component/index.js
