+++ internal/docs/component/prop-string.js
