+++ internal/docs/component/prop-switcher.js
