+++ internal/docs/component/prop-type.js
