+++ internal/docs/component/props.js
